Folder "functions" renamed to "utilities"
