+++ bot/modules/utilities/activityLogger.js
