+++ bot/modules/utilities/getTime.js
