+++ bot/modules/utilities/presenceChanger.js
