+++ bot/modules/utilities/spamProtection.js
