+++ bot/modules/utilities/startingInfo.js
